backend/requirements.txt
django == 5.0
djangorestframework == 3.14.0
djangorestframework-simplejwt == 5.3.1
requests == 2.31.0
cloudinary == 1.37.0
pillow == 10.1.0
django-environ == 0.11.2
docker == 7.0.0
django-cors-headers == 4.3.1